requirements.txt
flask
jinja2
flask_sqlalchemy
flask-qrcode
argon2-cffi